% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttext.R
\name{analogies}
\alias{analogies}
\title{Get Analogies}
\usage{
analogies(model, word_triplets, k = 10L)
}
\arguments{
\item{model}{an object inheriting from \code{"fasttext"}.}

\item{word_triplets}{a character vector of length \code{}string giving the word.}

\item{k}{an integer giving the number of nearest neighbors to be returned.}
}
\value{
.
}
\description{
TODO
}
\examples{
\dontrun{
nearest_neighbors(model, c("berlin", "germany", "france"), k = 6L)
}
}
