% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttext.R
\name{evaluate.supervised_model}
\alias{evaluate.supervised_model}
\title{Evaluate the Model}
\usage{
\method{evaluate}{supervised_model}(model, file, k = 1L, threshold = 0)
}
\arguments{
\item{file}{a character string giving the location of the validation file.}

\item{k}{an integer giving the number of labels to be returned.}

\item{threshold}{a double giving the threshold.}
}
\description{
Evaluate the quality of the predictions.
             For the model evaluation precision and recall are used.
}
\examples{
\dontrun{
evaluate(model, file)
}
}
