% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttext.R
\name{ft_load}
\alias{ft_load}
\title{Load Model}
\usage{
ft_load(file)
}
\arguments{
\item{file}{a character string giving the name of the file
to be read in.}
}
\value{
an object inheriting from \code{"fasttext"}.
}
\description{
Load a previously saved model from file.
}
\examples{
\dontrun{
model <- ft_load("dbpedia.bin")
}
}
