% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttext.R
\name{ft_nearest_neighbors}
\alias{ft_nearest_neighbors}
\title{Get Nearest Neighbors}
\usage{
ft_nearest_neighbors(model, word, k = 10L)
}
\arguments{
\item{model}{an object inheriting from \code{"fasttext"}.}

\item{word}{a character string giving the word.}

\item{k}{an integer giving the number of nearest neighbors to be returned.}
}
\value{
.
}
\description{
TODO
}
\examples{
\dontrun{
ft_nearest_neighbors(model, "enviroment", k = 6L)
}
}
