% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttext.R
\name{ft_save}
\alias{ft_save}
\title{Write Model}
\usage{
ft_save(model, file, what = c("model", "vectors", "output"))
}
\arguments{
\item{model}{an object inheriting from \code{'fasttext'}.}

\item{file}{a character string giving the name of the file.}

\item{what}{a character string giving what should be saved.}
}
\description{
Write a previously saved model from file.
}
\examples{
\dontrun{
ft_save(model, "my_model.bin", what = "model")
}
}
