% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttext.R
\name{ft_sentence_vectors}
\alias{ft_sentence_vectors}
\title{Get Sentence Vectors}
\usage{
ft_sentence_vectors(model, sentences)
}
\arguments{
\item{model}{an object inheriting from \code{"fasttext"}.}

\item{sentences}{a character vector giving the sentences.}
}
\value{
a matrix containing the sentence vectors.
}
\description{
Obtain sentence vectors from a previously trained model.
}
\examples{
\dontrun{
ft_sentence_vectors(model, c("sentence", "vector"))
}
}
