% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttext.R
\name{ft_test}
\alias{ft_test}
\title{Evaluate the Model}
\usage{
ft_test(model, file, k = 1L, threshold = 0)
}
\arguments{
\item{model}{an object inheriting from \code{'fasttext'}.}

\item{file}{a character string giving the location of the validation file.}

\item{k}{an integer giving the number of labels to be returned.}

\item{threshold}{a double giving the threshold.}
}
\description{
Evaluate the quality of the predictions.
             For the model evaluation precision and recall are used.
}
\examples{
\dontrun{
ft_test(model, file)
}
}
