% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttext.R
\name{ft_words}
\alias{ft_words}
\title{Get Words}
\usage{
ft_words(model)
}
\arguments{
\item{model}{an object inheriting from \code{"fasttext"}.}
}
\value{
a character vector.
}
\description{
Obtain all the words from a previously trained model.
}
\examples{
\dontrun{
ft_words(model)
}
}
