% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttext.R
\name{get_word_vectors}
\alias{get_word_vectors}
\title{Get Word Vectors}
\usage{
get_word_vectors(model, words)
}
\arguments{
\item{model}{an object inheriting from \code{"fasttext"}.}

\item{words}{a character vector giving the words.}
}
\value{
a matrix containing the word vectors.
}
\description{
Obtain word vectors from a previously trained model.
}
\examples{
\dontrun{
get_word_vectors(model, c("word", "vector"))
}
}

