% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttext.R
\name{normalize}
\alias{normalize}
\title{Normalize}
\usage{
normalize(txt)
}
\arguments{
\item{txt}{a character vector to be normalized.}
}
\value{
a character vector.
}
\description{
Applies normalization to a given text.
}
\examples{
\dontrun{
normalize(some_text)
}
}

