% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttext_2R_interface.R
\name{predict_unknown_words}
\alias{predict_unknown_words}
\title{predict unknown words for the skipgram and cbow models}
\usage{
predict_unknown_words(skipgram_cbow_model_output = NULL,
  unknown_words_path = NULL, output_path = NULL, verbose = FALSE)
}
\arguments{
\item{skipgram_cbow_model_output}{the output.bin of the \emph{skipgram_cbow} function}

\item{unknown_words_path}{a character string specifying the path to the unknown words file}

\item{output_path}{a character string specifying the path to the output file}

\item{verbose}{either TRUE or FALSE. If TRUE information will be printed out}
}
\value{
a character string specifying the location of the saved data
}
\description{
predict unknown words for the skipgram and cbow models
}
\details{
the function will save to a pre-specified path (\emph{output_path}) the numeric vectors for the unknown words. In case of non-matches it returns zero-valued vectors
}
\examples{

# library(fastTextR)

# res = predict_unknown_words(skipgram_cbow_model_output = "/data_fasttext/model.bin", 

#                             unknown_words_path = "/data_fasttext/queries.txt", 

#                             output_path = "/data_fasttext/NEW_VEC")
}
\references{
https://github.com/facebookresearch/fastText

https://arxiv.org/abs/1607.04606

https://arxiv.org/abs/1607.01759
}
