% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttext.R
\name{read.fasttext}
\alias{read.fasttext}
\title{Read Model}
\usage{
read.fasttext(file)
}
\arguments{
\item{file}{a character string giving the name of the file
to be read in.}
}
\value{
an object inheriting from \code{"fasttext"}.
}
\description{
Read a previously saved model from file.
}
\examples{
\dontrun{
model <- read.fasttext( "dbpedia.bin" )
}
}

