% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttext.R
\name{save.fasttext}
\alias{save.fasttext}
\title{Save Model}
\usage{
save.fasttext(model, file)
}
\arguments{
\item{model}{an object inheriting from \code{"fasttext"}.}

\item{file}{a character string giving the name of the file.}
}
\description{
Save the model to a file.
}
\examples{
\dontrun{
save.fasttext(model = m, file = "data.model")
}
}

