% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttext_2R_interface.R
\name{skipgram_cbow}
\alias{skipgram_cbow}
\title{wrapper for the skipgram and cbow methods of the fasttext text classifier}
\usage{
skipgram_cbow(input_path = NULL, output_path = NULL,
  method = "skipgram", lr = 0.1, lrUpdateRate = 100, dim = 100,
  ws = 5, epoch = 5, minCount = 1, neg = 5, wordNgrams = 1,
  loss = "ns", bucket = 2e+06, minn = 0, maxn = 0, thread = 6,
  t = 1e-04, verbose = 2)
}
\arguments{
\item{input_path}{a character string specifying the path to the train text file}

\item{output_path}{a character string specifying the path to the output-file}

\item{method}{a string specifying the method. One of \emph{skipgram}, \emph{cbow}}

\item{lr}{a float number specifying the learning rate [default is 0.1]}

\item{lrUpdateRate}{a number specifying the rate of updates for the learning rate [default is 100]}

\item{dim}{a number specifying the size of the word vectors [default is 100]}

\item{ws}{a number specifying the size of the context window [default is 5]}

\item{epoch}{a value specifying the number of epochs [default is 5]}

\item{minCount}{a number specifying the minimal number of word occurences [default is 1]}

\item{neg}{a value specifying the number of negatives sampled [default is 5]}

\item{wordNgrams}{a number specifying the max length of word ngram [default is 1]}

\item{loss}{a character string specifying the loss function. One of \emph{ns (negative sampling)}, \emph{hs (hierarchical softmax)}, \emph{softmax} [default is 'ns']}

\item{bucket}{a value specifying the number of buckets [default is 2000000]}

\item{minn}{a number specifying the min length of char ngram [default is 0]}

\item{maxn}{a number specifying the max length of char ngram [default is 0]}

\item{thread}{a value specifying the of threads [default is 6]}

\item{t}{a float number specifying the sampling threshold [default is 0.0001]}

\item{verbose}{a number (between 0 and 2) specifying the verbosity level [default is 2]}
}
\value{
a character string specifying the location of the saved data and the number of the word vectors
}
\description{
wrapper for the skipgram and cbow methods of the fasttext text classifier
}
\details{
the function will save a model.bin and the word vectors to a pre-specified path (\emph{output_path})
}
\examples{

# library(fastTextR)

# res = skipgram_cbow(input_path = "/data_fasttext/out_test_file.txt", 

#                     output_path = "/data_fasttext/model", method = "skipgram")
}
\references{
https://github.com/facebookresearch/fastText

https://arxiv.org/abs/1607.04606

https://arxiv.org/abs/1607.01759
}
