% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttext.R
\name{write.fasttext}
\alias{write.fasttext}
\title{Write Model}
\usage{
write.fasttext(object, file, what = c("model", "vectors", "output"))
}
\arguments{
\item{object}{an object inheriting from \code{'fasttext'}.}

\item{file}{a character string giving the name of the file.}

\item{what}{a character string giving what should be saved.}
}
\description{
Write a previously saved model from file.
}
\examples{
\dontrun{
write.fasttext(model, "my_model.bin", what = "model")
}
}
