% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttext.R
\name{fasttext}
\alias{fasttext}
\title{Create a New \code{FastText} Model}
\usage{
fasttext()
}
\description{
Create a new \code{FastText} model. The available methods
 are the same as the package functions but with out the prefix \code{"ft_"}
 and without the need to provide the model.
}
\examples{
ft <- fasttext()
}
