% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttext.R
\name{fasttext}
\alias{fasttext}
\title{Train a Model}
\usage{
fasttext(input, method = c("supervised", "cbow", "skipgram"),
  control = ft.control())
}
\arguments{
\item{input}{a character string giving the location of the input file.}

\item{method}{a character string giving the method, possible values are 
\code{'supervised'}, \code{'cbow'} and \code{'skipgram'}.}

\item{control}{a list giving the control variables, for more information
see \code{\link{ft.control}}.}
}
\description{
Train a new word representation model or supervised 
             classification model.
}
\examples{
\dontrun{
model <- fasttext("my_data.txt", method="supervised", 
                  control = ft.control(nthreads = 1L))
}
}

